Fix global methods after upgrading Babushka
The new version of Babushka loads all files in an anonymous module
(Kernel.load <file>, true) so the global namespace won't be polluted by
methods/variables defined "globally", ie outside a dep.

That means logic inside a dep cannot reference a "global" method, only
"global" constants. So all methods is moved inside a dep or if shared
between several deps in a file, let it be a constant outside the dep(s).

This happened in the upgrade of Babushka from v0.9.1 to v0.9.2.